+++ CHANGELOG.md
@@ -8,6 +8,11 @@ v2.9.0
 * Added support for value objects, a preview feature described by JEP 401 and JEP 539.
 * Added support for creating lambda functions.
 
+v2.8.3
+------
+* Fixed handling of NaN comparisons for the variants that don't act on a label. For example,
+  `lt(Object)` and `ifLt(Object, Runnable)`.
+
 v2.8.2 (2026-08-07)
 ------
 * Fixed code generation of empty catch blocks (no exception table entry should be generated).